refactor: worked on reviews
